+++ utils/index.ts
@@ -13,7 +13,6 @@
 
 export * from "./assert-never";
 export * from "./chunk-array";
-export * from "./chunk-call";
 export * from "./generate-random-id";
 export * from "./set-header-if-not-set";
 export * from "./set-not-enumerable-property";